Create views/bulma.blade.php Template
Bulma is becoming an extremely popular framework, especially over the past 6 months or so. I think its popularity is similar to Foundation6, and so I think it is worthy of a pre-built template. I followed the default breadcrumbs template defined here: http://bulma.io/documentation/components/breadcrumb/ so that anyone already using the Bulma framework will be able to instantly use the Laravel Breadcrumbs package without any customization. Of course it can be customized further by adding a class or two onto the main `<nav class="breadcrumb">` element. Just a note, the Bulma framework requires that even the last element be an <a href="#"> even though it isn't clickable, which is why I structured it the way I did.